+++ NEWS.md
@@ -1,5 +1,5 @@
 # pmml 2.0.0.9000
-
+* Application version in PMML Header corresponds to pmml package version.
 
 # pmml 2.0.0